% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{calc_nspp}
\alias{calc_nspp}
\title{Calculate number of species (nspp)}
\usage{
calc_nspp(con, tif, col_grps, group, tbl_spp_cells, col_spp,
  spp_prob_threshold = 0.5)
}
\arguments{
\item{con}{database connection}

\item{tif}{raster tif output}

\item{col_grps}{column in species table with taxonomic groupings}

\item{group}{taxonomic group}

\item{tbl_spp_cells}{species_cells table with modeling period (ie now or 2100)}

\item{col_spp}{column in species_cells matching spp.SPECIESID}

\item{spp_prob_threshold}{probability threshold (default=0.5) to count as present}
}
\value{
True of completed, False if not
}
\description{
Calculate number of species (nspp)
}
