% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{csv_to_db}
\alias{csv_to_db}
\title{load csv data into database}
\usage{
csv_to_db(con, pfx = "/Users/bbest/docker_pg_data/aquamaps_csv",
  pfx_db = "/data/aquamaps_csv", tbl_csv = list(spp =
  "speciesoccursum_ver0816c.csv", obs_cells =
  "occurrencecells_ver0816c.csv", env_cells = "hcaf_v6.csv", spp_cells =
  "hcaf_species_native_ver0816c.csv"), redo = T)
}
\arguments{
\item{con}{database connection}

\item{pfx}{prefix for locally available data}

\item{pfx_db}{prefix on database server}

\item{tbl_csv}{named list of database table names = csv's to import}
}
\value{
TRUE
}
\description{
load csv data into database
}
\examples{
\dontrun{
csv_to_db(con)
}
}
