% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{df_to_raster}
\alias{df_to_raster}
\title{data frame to global half-degree raster}
\usage{
df_to_raster(df, col_value = "value", tif = NULL,
  col_cellid = "cellid", cols_lonlat = NULL)
}
\arguments{
\item{df}{data frame containing col_cellid OR cols_lonlat and value column}

\item{col_value}{column of value to convert to raster}

\item{tif}{optional output tif to write}

\item{col_cellid}{column of col_cellid in data frame, default "cellid", NULL if using cols_lonlat}

\item{cols_lonlat}{columns of cols_lonlat in data frame, eg `c("CenterLat","CenterLong")`}
}
\value{
raster of given value in data.frame by location given for half-degree global raster
}
\description{
Convert data.frame to global 0.5 deg raster
}
