% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_db_con}
\alias{get_db_con}
\title{get database connection}
\usage{
get_db_con(host = "localhost", port = 5432, dbname = "am",
  user = "admin", password = "CHANGEME!")
}
\arguments{
\item{host}{defaults to "localhost"}

\item{port}{defaults to 5432}

\item{dbname}{defaults to "am"}

\item{user}{defaults to "admin"}

\item{password}{defaults to "CHANGEME!"}
}
\value{
DBI::dbConnect()
}
\description{
get database connection
}
\examples{
con <- get_db_con(password=readLines("~/private/pg_pass"))
}
