% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{get_iucn}
\alias{get_iucn}
\title{get IUCN Red List information}
\usage{
get_iucn(genus_species, rl_token)
}
\arguments{
\item{rl_token}{RedList token, issued <http://apiv3.iucnredlist.org/api/v3/token>}
}
\value{
data frame of: category, criteria, population_trend, published_year
}
\description{
Return data.frame of extinction risk using IUCN RedList web service
}
\examples{
rl_token <- readLines("~/private/iucn_api_key")
rl_search("Megaptera novaeangliae", key=rl_token)
}
\seealso{
\code{\link{rredlist::rl_search}}
}
