% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmbi_stack}
\alias{gmbi_stack}
\title{Global marine biodiversity indicators from AquaMaps and IUCN}
\format{A \code{\link[raster]{stack}} with indicator (\code{nspp_*} or
  \code{rls_\*}) and taxonomic group or taxa (\code{\*_all}).}
\source{
\url{https://github.com/marinebon/bbnj}

\url{https://aquamaps.org}

\url{https://iucnredlist.org}
}
\usage{
gmbi_stack
}
\description{
Raster stack containing global half-degree rasters with species richness
(\code{nspp_\*}) and Red List Sum (\code{rls_\*}) for current distributions
and 2100 based on IPCC A2 emissions scenario.
}
\keyword{datasets}
