% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{group_to_grp}
\alias{group_to_grp}
\title{convert proper Group to grp abbreviation for file naming}
\usage{
group_to_grp(group)
}
\arguments{
\item{group}{proper name of taxonomic group}
}
\value{
lower case group, with " & " replaced by ".n." and spaces replaced by
  "."
}
\description{
convert proper Group to grp abbreviation for file naming
}
