% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{grp_to_group}
\alias{grp_to_group}
\title{convert grp abbreviation to proper Group for file naming}
\usage{
grp_to_group(grp)
}
\arguments{
\item{grp}{taxonomic group abbreviation}
}
\value{
title case group, with ".n." replaced by " & ", and "." replaced by
  spaces
}
\description{
convert grp abbreviation to proper Group for file naming
}
