% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_na}
\alias{raster_na}
\title{get global half-degree raster with NAs}
\usage{
raster_na()
}
\value{
global half-degree raster populated with NAs
}
\description{
get global half-degree raster with NAs
}
