% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spp}
\alias{spp}
\title{AquaMaps species list}
\format{A data.frame with species information.}
\source{
\url{https://github.com/marinebon/bbnj}

\url{https://aquamaps.org}
}
\usage{
spp
}
\description{
Table of AquaMaps species
}
\keyword{datasets}
