% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spp_iucn}
\alias{spp_iucn}
\title{IUCN species information}
\format{A data.frame with species information.}
\source{
\url{https://github.com/marinebon/bbnj}

\url{https://iucnredlist.org}
}
\usage{
spp_iucn
}
\description{
Table of IUCN species information
}
\keyword{datasets}
